% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closure.R
\name{new_closure}
\alias{new_closure}
\title{A foolproof construction of closure.}
\usage{
new_closure(members, parent_env = parent.frame(), keep_promise = character(0))
}
\description{
A foolproof construction of closure.
}
